`storage` stuff moved from `model` to `core`
