+++ server/server/src/spaces.ts
@@ -13,8 +13,8 @@
 // limitations under the License.
 //
 
-import { AnyLayout, Class, Doc, Ref, Tx, StringProperty } from '@anticrm/model'
-import core, { CreateTx, DeleteTx, PushTx, Space, SpaceUser, UpdateTx } from '@anticrm/core'
+import { AnyLayout, Class, Doc, Ref, StringProperty } from '@anticrm/model'
+import core, { CreateTx, DeleteTx, PushTx, Space, SpaceUser, UpdateTx, Tx } from '@anticrm/core'
 import { Client } from './server'
 import { WorkspaceProtocol } from './workspace'
